Switching bionic to new branch
We are undergoing a new strategy, instead of starting with aosp base we are going to try and merge google over it OG style and then resolve the conflicts...bionic is repo #2 thats doing it. In case this is failed though, we have the jb2 branches to fall back on. When the merges are complete current jb2 branches will be deleted and replaced by mr1. ;)